CROW-27 Add events abstract class to optimize services
